fix(node): wrap ClickHouse responseHook calls in try-catch to prevent orphaned spans (#15966)
