+++ packages/node/src/integrations/tracing/clickhouse/patch.ts
@@ -1,19 +1,9 @@
-import type { Span, Tracer } from '@opentelemetry/api';
+import type { Tracer } from '@opentelemetry/api';
 import { context, SpanKind, SpanStatusCode, trace } from '@opentelemetry/api';
 import type { InstrumentationBase } from '@opentelemetry/instrumentation';
 import { SEMANTIC_ATTRIBUTE_SENTRY_OP } from '@sentry/core';
 import type { ClickHouseInstrumentationConfig } from './types';
-
-interface ClickHouseSummary {
-  [key: string]: unknown;
-  elapsed_ns?: string;
-  read_bytes?: string;
-  read_rows?: string;
-  result_bytes?: string;
-  result_rows?: string;
-  written_bytes?: string;
-  written_rows?: string;
-}
+import { addExecutionStats, extractOperation, extractSummary, sanitizeQueryText } from './utils';
 
 export interface ClickHouseModuleExports {
   ClickHouseClient: unknown;
@@ -35,132 +25,14 @@ const SEMATTRS_DB_NAME = 'db.name';
 const SEMATTRS_NET_PEER_NAME = 'net.peer.name';
 const SEMATTRS_NET_PEER_PORT = 'net.peer.port';
 
-// ClickHouse execution statistics attributes
-const SEMATTRS_CLICKHOUSE_READ_ROWS = 'clickhouse.read_rows';
-const SEMATTRS_CLICKHOUSE_READ_BYTES = 'clickhouse.read_bytes';
-const SEMATTRS_CLICKHOUSE_WRITTEN_ROWS = 'clickhouse.written_rows';
-const SEMATTRS_CLICKHOUSE_WRITTEN_BYTES = 'clickhouse.written_bytes';
-const SEMATTRS_CLICKHOUSE_RESULT_ROWS = 'clickhouse.result_rows';
-const SEMATTRS_CLICKHOUSE_RESULT_BYTES = 'clickhouse.result_bytes';
-const SEMATTRS_CLICKHOUSE_ELAPSED_NS = 'clickhouse.elapsed_ns';
-
-/**
- * Extracts the SQL operation (SELECT, INSERT, etc.) from query text.
- */
-function extractOperation(queryText: string): string | undefined {
-  const trimmed = queryText.trim();
-  const match = /^(?<op>\w+)/u.exec(trimmed);
-  return match?.groups?.op?.toUpperCase();
-}
-
-/**
- * Sanitizes and truncates query text for safe inclusion in spans.
- */
-function sanitizeQueryText(queryText: string, maxLength: number): string {
-  if (queryText.length <= maxLength) {
-    return queryText;
-  }
-  return `${queryText.substring(0, maxLength)}...`;
-}
-
-/**
- * Extracts ClickHouse summary from response headers.
- */
-function extractSummary(headers: Record<string, unknown>): ClickHouseSummary | undefined {
-  if (!headers) {
-    return undefined;
-  }
-
-  const summary = headers['x-clickhouse-summary'] as string | undefined;
-  if (summary && typeof summary === 'string') {
-    try {
-      return JSON.parse(summary);
-    } catch {
-      return undefined;
-    }
-  }
-
-  if ('read_rows' in headers || 'result_rows' in headers || 'elapsed_ns' in headers) {
-    return headers;
-  }
-
-  return undefined;
-}
-
-/**
- * Adds ClickHouse execution statistics to span attributes.
- */
-function addExecutionStats(span: Span, summary: ClickHouseSummary): void {
-  if (!summary) {
-    return;
-  }
-
-  try {
-    if (summary.read_rows !== undefined) {
-      const readRows = parseInt(summary.read_rows, 10);
-      if (!isNaN(readRows)) {
-        span.setAttribute(SEMATTRS_CLICKHOUSE_READ_ROWS, readRows);
-      }
-    }
-
-    if (summary.read_bytes !== undefined) {
-      const readBytes = parseInt(summary.read_bytes, 10);
-      if (!isNaN(readBytes)) {
-        span.setAttribute(SEMATTRS_CLICKHOUSE_READ_BYTES, readBytes);
-      }
-    }
-
-    if (summary.written_rows !== undefined) {
-      const writtenRows = parseInt(summary.written_rows, 10);
-      if (!isNaN(writtenRows)) {
-        span.setAttribute(SEMATTRS_CLICKHOUSE_WRITTEN_ROWS, writtenRows);
-      }
-    }
-
-    if (summary.written_bytes !== undefined) {
-      const writtenBytes = parseInt(summary.written_bytes, 10);
-      if (!isNaN(writtenBytes)) {
-        span.setAttribute(SEMATTRS_CLICKHOUSE_WRITTEN_BYTES, writtenBytes);
-      }
-    }
-
-    if (summary.result_rows !== undefined) {
-      const resultRows = parseInt(summary.result_rows, 10);
-      if (!isNaN(resultRows)) {
-        span.setAttribute(SEMATTRS_CLICKHOUSE_RESULT_ROWS, resultRows);
-      }
-    }
-
-    if (summary.result_bytes !== undefined) {
-      const resultBytes = parseInt(summary.result_bytes, 10);
-      if (!isNaN(resultBytes)) {
-        span.setAttribute(SEMATTRS_CLICKHOUSE_RESULT_BYTES, resultBytes);
-      }
-    }
-
-    if (summary.elapsed_ns !== undefined) {
-      const elapsedNs = parseInt(summary.elapsed_ns, 10);
-      if (!isNaN(elapsedNs)) {
-        span.setAttribute(SEMATTRS_CLICKHOUSE_ELAPSED_NS, elapsedNs);
-      }
-    }
-  } catch {
-    // Silently ignore errors in stats extraction
-  }
-}
-
 // Type definitions for ClickHouse client internals
 interface ClickHouseClientInstance {
   query: unknown;
   insert: unknown;
   exec: unknown;
   command: unknown;
-  connection_params?: {
-    url?: string;
-  };
-  options?: {
-    url?: string;
-  };
+  connection_params?: { url?: string };
+  options?: { url?: string };
 }
 
 interface ClickHouseQueryParams {
@@ -191,19 +63,18 @@ export function patchClickHouseClient(
   const { wrap, tracer, getConfig, isEnabled } = options;
   const ClickHouseClient = moduleExports.ClickHouseClient;
 
-    if (!ClickHouseClient || typeof ClickHouseClient !== 'function' || !('prototype' in ClickHouseClient)) {
+  if (!ClickHouseClient || typeof ClickHouseClient !== 'function' || !('prototype' in ClickHouseClient)) {
     return moduleExports;
   }
 
-    const ClickHouseClientCtor = ClickHouseClient as new () => {
-      query: unknown;
-      insert: unknown;
-      exec: unknown;
-      command: unknown;
-    };
-    const prototype = ClickHouseClientCtor.prototype;
+  const ClickHouseClientCtor = ClickHouseClient as new () => {
+    query: unknown;
+    insert: unknown;
+    exec: unknown;
+    command: unknown;
+  };
+  const prototype = ClickHouseClientCtor.prototype;
 
-  // Helper to patch standard query methods
   const patchGeneric = (methodName: string): void => {
     wrap(
       prototype,
@@ -216,7 +87,6 @@ export function patchClickHouseClient(
     );
   };
 
-  // Helper to patch insert specifically
   const patchInsert = (): void => {
     wrap(
       prototype,
@@ -226,7 +96,6 @@ export function patchClickHouseClient(
         const table = params.table || '<unknown>';
         const format = params.format || 'JSONCompactEachRow';
         let statement = `INSERT INTO ${table}`;
-
         if (params.columns) {
           if (Array.isArray(params.columns)) {
             statement += ` (${params.columns.join(', ')})`;
@@ -235,11 +104,7 @@ export function patchClickHouseClient(
           }
         }
         statement += ` FORMAT ${format}`;
-
-        return {
-          queryText: statement,
-          operation: 'INSERT', // Explicitly force INSERT operation
-        };
+        return { queryText: statement, operation: 'INSERT' };
       }),
     );
   };
@@ -251,12 +116,6 @@ export function patchClickHouseClient(
 
   return moduleExports;
 }
-
-/**
- * A generic patch handler factory that handles the boilerplate
- * of span creation, context wrapping, execution, and error handling.
- */
-// patch.ts (Partial update - replace the createPatchHandler function)
 
 function createPatchHandler(
   methodName: string,
@@ -275,7 +134,7 @@ function createPatchHandler(
       let extraction;
       try {
         extraction = attributesExtractor(args);
-      } catch (e) {
+      } catch {
         extraction = { queryText: '' };
       }
 
@@ -295,22 +154,16 @@ function createPatchHandler(
       if (config.dbName) {
         span.setAttribute(SEMATTRS_DB_NAME, config.dbName);
       }
-
       if (config.captureQueryText !== false && queryText) {
         const maxLength = config.maxQueryLength || 1000;
         span.setAttribute(SEMATTRS_DB_STATEMENT, sanitizeQueryText(queryText, maxLength));
       }
-
-      // Connection Attributes Logic:
-      // 1. Prefer explicit config
       if (config.peerName) {
         span.setAttribute(SEMATTRS_NET_PEER_NAME, config.peerName);
       }
       if (config.peerPort) {
         span.setAttribute(SEMATTRS_NET_PEER_PORT, config.peerPort);
       }
-
-      // 2. Fallback to auto-discovery if attributes are missing
       if (!config.peerName || !config.peerPort) {
         try {
           const clientConfig = this.connection_params || this.options;
@@ -320,7 +173,6 @@ function createPatchHandler(
               span.setAttribute(SEMATTRS_NET_PEER_NAME, url.hostname);
             }
             if (!config.peerPort) {
-              // Ensure port is stored as a number
               span.setAttribute(SEMATTRS_NET_PEER_PORT, parseInt(url.port, 10) || 8123);
             }
           }
@@ -334,15 +186,18 @@ function createPatchHandler(
           if (config.captureExecutionStats !== false && response) {
             const headers = response.response_headers || response.headers;
             if (headers) {
-                const summary = extractSummary(headers);
-                if (summary) {
-                  addExecutionStats(span, summary);
-                }
+              const summary = extractSummary(headers);
+              if (summary) {
+                addExecutionStats(span, summary);
+              }
             }
           }
-          // Call responseHook with the actual response
           if (config.responseHook) {
-            config.responseHook(span, response);
+            try {
+              config.responseHook(span, response);
+            } catch {
+              // Ignore errors from user-provided hooks
+            }
           }
           span.setStatus({ code: SpanStatusCode.OK });
           span.end();
@@ -350,12 +205,15 @@ function createPatchHandler(
         };
 
         const onError = (error: Error): never => {
-          // Call responseHook to ensure sentry.origin is set for error cases
           if (config.responseHook) {
-            config.responseHook(span, undefined);
+            try {
+              config.responseHook(span, undefined);
+            } catch {
+              // Ignore errors from user-provided hooks
+            }
           }
           span.recordException(error);
-          span.setStatus({code: SpanStatusCode.ERROR, message: error.message});
+          span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
           span.end();
           throw error;
         };
